==> 03 Scripts/1-data-wrangling-and-subsetting.ipynb <==
# Data Wrangling

### Importing libraries

import pandas as pd
import numpy as np
import os

### Creating Dataframes from: orders, products and departments

# Establish Data path
data_path = r'C:\Users\gerar\CareerFoundry Python\Instacart Basket Analysis\02 Data'

# orders.csv into df_orders
df_orders = pd.read_csv(os.path.join(data_path, 'Original Data', 'orders.csv'))


# products.csv into df_products
df_products = pd.read_csv(os.path.join(data_path, 'Original Data', 'products.csv'))


# departments.csv into df_departments
df_departments = pd.read_csv(os.path.join(data_path, 'Original Data', 'departments.csv'))

df_orders.shape

df_products.shape

df_departments.shape

### Data Wrangling

#dropping 'eval_set' column and update df_orders
df_orders = df_orders.drop(columns = ['eval_set'])

# renaming a column title in df_orders.   From 'order_dow'  to  'order_day_of_the_week' 
df_orders.rename(columns = {'order_dow': 'order_day_of_the_week'}, inplace = True)

# changing 'order_id' from df_orders into a string type
df_orders['order_id'] = df_orders['order_id'].astype('str')

# transposing df_departments into df_departments_T
df_departments_t = df_departments.T

# getting the header from the transposed table using .iloc
departments_header = df_departments_t.iloc[0,:]

# getting rid of the first row of df_departments. Create df_departments_t_new
df_departments_t_new = df_departments_t[1:]

# reassigning column headers to table
df_departments_t_new.columns = departments_header

# turning df_departments_t_new into a data dictionary
departments_dictionary = df_departments_t_new.to_dict('index')

departments_dictionary

### Subsetting

# Creates a subset 'df_snacks' from 'df_products' by selecting rows that fulfill the condition 'department_id'==19

df_snacks = df_products[df_products['department_id']==19]

# Task 

### 2. Changing a datatype

# changes 'user_id' column into a string type
df_orders['user_id'] = df_orders['user_id'].astype('str')

df_orders.dtypes

### 3. Changing a column name

# Renames 'order_number' into 'customer_order_count'. It specifices if this is the first, second ...etc of this customer
df_orders.rename(columns = {'order_number': 'customer_order_count'}, inplace = True)

df_orders.head(0)

### 4. Busiest hour for placing orders?

# Counts grouping by 'order_hour_of_day'
df_orders['order_hour_of_day'].value_counts(dropna = False)

##### The busiest hour for orders is 10 am

### 5. Meaning of "4" in department_id? 

print(departments_dictionary.get('4'))

departments_dictionary

### 6. Subset of breakfast items

# creates dataframe from df_products for rows with department_id = 14, "breakfast"
df_breakfast_items = df_products[df_products['department_id']==14]

df_breakfast_items

### 7. Subset that includes: alcohol, deli, beverages and meat/seafood

# creates dataframe out of alcohol,deli,beverages and meat/seafood 
df_dinner_party = df_products.loc[df_products['department_id'].isin([5,7,12,20])]

### 8. Row count of df_dinner_party

# counts rows in df_dinner_party
print("The dinner party subset has: ", len(df_dinner_party), "rows.")

### 9. Information about user with user_id = 1

# the user_id 1 will be represented as df_customer1 
df_customer1 = df_orders[df_orders['user_id']=='1']

df_customer1

### 10. Basic Stats on customer 1

# Provides information about the customer with user_id '1'
df_customer1.describe()

I don't see anything suspicious with user_id == 1. Further analysis is needed 

### 11. Exporting Data

# Exports df_orders as orders_wrangled.csv
df_orders.to_csv(os.path.join(data_path, 'Prepared Data', 'orders_wrangled.csv'))

# Exports df_departments_t_new as departments_wrangled.csv
df_departments_t_new.to_csv(os.path.join(data_path, 'Prepared Data', 'departments_wrangled.csv'))
