==> 03 Scripts/2-data-consistency-checks.ipynb <==
### Importing Libraries

import pandas as pd
import numpy as np
import os

### Importing Data

# declares data path 
data_path = r'C:\Users\gerar\CareerFoundry Python\Instacart Basket Analysis\02 Data'

# imports products.csv into df_products
df_products = pd.read_csv(os.path.join(data_path, 'Original Data', 'products.csv'))

# imports orders_wrangled.csv into df_orders
df_orders_clean = pd.read_csv(os.path.join(data_path, 'Prepared Data', 'orders_wrangled.csv'))

# check shape of imported dataframes 
print('This is the shape of the imported orders table:', df_orders_clean.shape)
print('This is the shape of the imported products table', df_products_clean.shape)

### Consistency Checks

# checks how many null values are in a dataframe
df_products.isnull().sum()

# creates a subset without rows that have null values in the product_name column
df_products_clean = df_products[df_products['product_name'].isnull()== False]

# checks for duplicate values and creates a safety dataframe for these duplicates
df_duplicates = df_products_clean[df_products_clean.duplicated()]

df_duplicates

# removes duplicates from df_products_clean and creates a new dataframe
df_products_clean_no_duplicates = df_products_clean.drop_duplicates()

# amount of rows df_products_clean
df_products_clean.shape

# Task

### 2. Describe df_orders to discover columns with unexpected values

df_orders_clean.describe()

2. I don't see anything that seems irregular, it should be further investigated

### 3. Checking for mixed-type data in the orders dataframe

for col in df_orders_clean.columns.tolist():
  weird = (df_orders_clean[[col]].map(type) != df_orders_clean[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_orders_clean[weird]) > 0:
    print (col)

### 4. Comment if there was mixed-type data in the orders dataframe

No mixed-type data was found. This code above does not print any column name. 

### 5. Checking for missing values in the orders dataframe

# check df_orders_clean for missing values
df_orders_clean.isnull().sum()

# creates temporal subset to explore those rows

df_orders_clean[df_orders_clean['days_since_prior_order'].isnull() == True]

### 6. Addressing missing values

These values represent a customer ordering for the first time. This is why there is no prior days.

Therefore, I dont think they should be removed. 

### 7. Checking for duplicate values

# checks duplicate values
df_orders_clean[df_orders_clean.duplicated()]

No duplicate values found in the orders dataframe.

### 8. Adressing Duplicate values

No duplicates in df_orders_clean

# creating df_orders_clean_no_duplicates for naming consistency
df_orders_clean_no_duplicates = df_orders_clean

### 9. Exporting Data

# Exporting df_orders_clean and df_products_no_duplicates as file_consistent.csv

df_orders_clean_no_duplicates.to_csv(os.path.join(data_path, 'Prepared Data', 'orders_consistent.csv'))

df_products_clean_no_duplicates.to_csv(os.path.join(data_path, 'Prepared Data', 'products_consistent.csv'))

# check shape of exported dataframes 
print('This is the shape of the consistent orders table:', df_orders_clean_no_duplicates.shape)
print('This is the shape of the consistent products table', df_products_clean_no_duplicates.shape)
