==> 03 Scripts/3-combining-exporting-data.ipynb <==
### Importing Libraries

import pandas as pd
import numpy as np
import os

# setting up data path as a raw string
data_path = r'C:\Users\gerar\CareerFoundry Python\Instacart Basket Analysis\02 Data'

### Importing files

# importing pickle file as df_orders_products_combined
df_orders_products_combined = pd.read_pickle(os.path.join(data_path, 'Prepared Data', 'orders_products_combined.pkl'))

df_products_consistent = pd.read_csv(os.path.join(data_path, 'Prepared Data', 'products_consistent.csv'), index_col = False)

### 4. Checking shape of imported files

# shape 
df_orders_products_combined.shape

# shape
df_products_consistent.shape

### 5. Combining data frames 

# Checking keys of both dataframes
df_products_consistent.head(1)

# the last column "_merge" is causing an error when merging, it needs to be deleted
df_orders_products_combined

# getting rid of '_merge' column to avoid ValueError. 
df_orders_products_combined = df_orders_products_combined.drop('_merge', axis=1)

# merging df_products_consistent and df_orders_products_combined
df_orders_products_merge = df_products_consistent.merge(df_orders_products_combined, on = ['product_id'], indicator = True ) 

df_orders_products_merge

### 6. Confirming merge results

# value counts of merge flag
df_orders_products_merge['_merge'].value_counts()

### 7. Exporting merged dataframe

df_orders_products_merge.to_pickle(os.path.join(data_path, 'Prepared Data', 'orders_products_merge.pkl'))

# checking shape of exported dataframe
df_orders_products_merge.shape
