==> 03 Scripts/4-deriving-new-variables.ipynb <==
### Importing Libraries

import pandas as pd
import numpy as np
import os

### Importing Data

# establishes data path
data_path = r'C:\Users\gerar\CareerFoundry Python\Instacart Basket Analysis\02 Data'

# imports pickle file into a panda dataframe 
df_orders_products_merge = pd.read_pickle(os.path.join(data_path, 'Prepared Data', 'orders_products_merge.pkl'))

# creates a subset for safe experimenting
df = df_orders_products_merge[:1000000]

### Defining a function

def price_label(row):
    if row['prices'] <= 5:
        return 'Low-range product'
    elif (row['prices'] > 5) and (row ['prices'] <= 15):
        return 'Mid-range product'
    elif row['prices'] > 15: 
        return 'High range'
    else: return 'Not enough data'

### Applying price label function

# uses price label function and creates a price_range column in the df
df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts()

df['prices'].max()

df.loc[df['prices']> 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts()

df_orders_products_merge.loc[df_orders_products_merge['prices']<= 5, 'price_range_loc'] = 'Low-range product'

df_orders_products_merge.loc[(df_orders_products_merge['prices'] > 5) & (df_orders_products_merge['prices'] <= 15), 'price_range_loc'] = 'Mid_range_price'

df_orders_products_merge.loc[df_orders_products_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_orders_products_merge['price_range_loc'].value_counts()

#### For loop example

for x in range(30,45):
    print('My age is %d' %(x))

df_orders_products_merge['order_day_of_the_week'].value_counts(dropna = False)

result = []

for value in df_orders_products_merge['order_day_of_the_week']:
    if value == 0:
        result.append('Busiest day')
    elif value == 4:
        result.append('Least busy')
    else:
        result.append('Regularly busy')

df_orders_products_merge['busiest_day'] = result

df_orders_products_merge['busiest_day'].value_counts(dropna = False)

# Task

### 2. Modifying busiest day column

# resets the "result" list. 
# modifies "for" loop with the new conditions. New logic: top 2 and bottom 2. The rest stays the same.

result = []

for value in df_orders_products_merge['order_day_of_the_week']:
    if (value == 0) or (value == 1):
        result.append('Busiest days')
    elif (value == 4 or value == 3):
        result.append('Least busy days')
    else:
        result.append('Regularly busy')

# assigns the new result list into a new "busiest_day_new" column 

df_orders_products_merge['busiest_days_new'] = result

# checks the value counts of the modified labels in the "busiest day column"
df_orders_products_merge['busiest_days_new'].value_counts(dropna = False)

# checks the value counts of the original column for comparison
df_orders_products_merge['busiest_day'].value_counts(dropna = False)

### 3. Observations of busiest_days_new  column

If we compare both values we can see that the old column that the top 2 values are almost as high as the regularly busy days. Using simple calculations we can have the value of Top 1, Top 2, Bottom 1 and Bottom 2 days. 

print('This is the top 1 value: ',  6204182)
print('This is the top 2 value: ',  11864412-6204182)
print('This is the bottom 1 value: ', 3783802)
print('This is the bottom 2 value: ', 7624336-3783802)

### 4. Busiest hour ranges

df_orders_products_merge.head(1)

# uses values counts of all the hours in to visualiye possible hour ranges 
df_orders_products_merge['order_hour_of_day'].value_counts(dropna =False)


# declares 2 lists which represent top and bottom periods. 'Most orders' and 'Fewest Orders'.
most_orders = [10, 11, 14, 15, 13, 12]
fewest_orders = [0, 1, 5, 2, 4, 3]

# declares empty list that will fill the 'busiest_period_of_day' column
busiest_period_list = []

# for loop that fills the empty list 
for value in df_orders_products_merge['order_hour_of_day']:  
    if value in most_orders:
        busiest_period_list.append('Most orders')
    elif value in fewest_orders:
        busiest_period_list.append('Fewest orders')
    else:
        busiest_period_list.append('Average orders')


# creates a column inside the df_orders_products_merge dataframe. Column name is 'busiest_period_of_day'.
# assigns the values of busiest_period_list into this column

df_orders_products_merge['busiest_period_of_day'] = busiest_period_list

# checks the results
df_orders_products_merge.head(1)

### 5. Print the frequency for this new column

# print the counts of those periods
print(df_orders_products_merge['busiest_period_of_day'].value_counts(dropna = False))

### 7. Exporting dataframe

df_orders_products_merge.to_pickle(os.path.join(data_path, 'Prepared Data', 'orders_products_merge_busy.pkl'))

# checks shape of exported datafram
df_orders_products_merge.shape
