==> 03 Scripts/6-adding-extra-data.ipynb <==
# Task

### 3. Importing Modules

import pandas as pd
import numpy as np
import os 
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

### Importing Data

# declare data path as a raw string and import customer into df_customer

data_path = r'C:\Users\gerar\CareerFoundry Python\Instacart Basket Analysis'
df_customers = pd.read_csv(os.path.join(data_path, '02 Data', 'Original Data', 'customers.csv'))
df_orders_products_merge = pd.read_pickle(os.path.join(data_path, '02 Data', 'Prepared Data', 'orders_products_merge_group.pkl'))

# check shapes

print(f' These are the dimensions of the customers dataframe: {df_customers.shape}')
print(f' These are the dimensions of the orders-products merge dataframe: {df_orders_products_merge.shape}')

#check headers

df_customers.head(1)

### 4. Data Wrangling

# Change the name of columns 
# write a list with the new column names

headers_customers = ['user_id', 'first_name', 'last_name', 'gender', 'state',
                     'age', 'date_joined', 'number_of_dependants', 'marital_status', 'income']
# test
len(headers_customers) == len(df_customers.columns)

df_customers.columns = headers_customers

# test
df_customers.head(1)

### Summary: Data Wrangling
Only the column names were changed 

### 5. Data consistency

# exploratory analysis of variables
df_customers.describe(include = 'all')

#checking data types 
df_customers.dtypes

# change user_id into object type
df_customers['user_id'] = df_customers['user_id'].astype(str)

# find how many null/NaN values in the dataframe
df_customers.isna().sum()

# create a slice with these customers as safety
df_customers_no_name= df_customers[df_customers['first_name'].isna()]

#check for duplicates 
df_customers[df_customers.duplicated()]

# check for mixed type data 
for col in df_customers.columns.tolist():
  weird = (df_customers[[col]].map(type) != df_customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_customers[weird]) > 0:
    print (col)

# check if the nan values are causing mixed data type in first_name of something else
df_customers_with_names = df_customers[df_customers['first_name'].notna()]

# check for mixed type data 
for col in df_customers_with_names.columns.tolist():
  weird = (df_customers_with_names[[col]].map(type) != df_customers_with_names[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_customers_with_names[weird]) > 0:
    print (col)

### Summary: Data Consistency


-user_id is now an object type 

-nan rows in first_name were spotted (need further advice on how to treat them)

-no duplicates were found

-mixed type data in first_name is caused by nan entries


### 5. Data combination

# display all column names
pd.options.display.max_columns = None

# check headers of both dataframes
df_customers.head(1)

df_orders_products_merge.head(1)

# drop "_merge" column Indicator
df_orders_products_merge = df_orders_products_merge.drop('_merge', axis=1)

#check compatible datatypes of key 
print(df_orders_products_merge['user_id'].dtype)

print(df_customers['user_id'].dtype)

# make df_orders_products_merge['user_id'] an object type 
df_orders_products_merge['user_id'] = df_orders_products_merge['user_id'].astype(str)

print(df_orders_products_merge['user_id'].dtype)

# merging tables 
df_orders_products_all = pd.merge(df_customers, df_orders_products_merge, on='user_id', how = 'outer', indicator = True)

### Summary: Data Combination

# shape
df_orders_products_all.shape

### Data export

df_orders_products_all.to_pickle(os.path.join(data_path, '02 Data', 'Prepared Data', 'orders_products_all.pkl'))

# shape
df_orders_products_all.shape
